Updated POD Signage
I've updated the POD signage with the interactive menu selection effect,
changing the menu variable on the players will allow the presenter to
stop and the start the signage.

Uploaded Proxima font the to the players, initially i had created PNG
using the Proxima font as i didnt know if you could upload new fonts to
the express player.

Tidied up transitions.